requirements.txt
streamlit>=1.31.0
google-generativeai>=0.8.0
Pillow>=10.0.0

